requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
libvirt-python==8.0.0

